- First pass through with with device tree enhancement merge.  Most of the mechanisms should   be in place but don't expect anything to quite work yet.
git-svn-id: svn://svn.coreboot.org/coreboot/trunk@1662 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/include/device/device.h
@@ -42,8 +42,8 @@ struct device {
 	device_t	next;		/* chain of all devices */
 
 	struct device_path path;
-	unsigned short	vendor;
-	unsigned short	device;
+	unsigned 	vendor;
+	unsigned 	device;
 	unsigned int	class;		/* 3 bytes: (base,sub,prog-if) */
 	unsigned int	hdr_type;	/* PCI header type */
 	unsigned int    enabled : 1;	/* set if we should enable the device */
